rename program-data to bpf-upgradable-state
