+++ tests/bpf-upgradable-state/Anchor.toml
@@ -1,5 +1,5 @@
 [programs.localnet]
-program_data = "Cum9tTyj5HwcEiAmhgaS7Bbj4UczCwsucrCkxRECzM4e"
+bpf_upgradable_state = "Cum9tTyj5HwcEiAmhgaS7Bbj4UczCwsucrCkxRECzM4e"
 
 [registry]
 url = "https://anchor.projectserum.com"
